Gen AI Projects/Multiple_PDF_Chat_app/requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
huggingface_hub
google-generativeai
langchain-community
langchain-google-genai
